Move Android specific stuff to util.cpp.
Tested: on OSX/Android.
