+++ API-Samples/spirv_assembly/spirv_assembly.cpp
@@ -195,7 +195,6 @@ int sample_main()
     init_device_extension_names(info);
     init_instance(info, sample_title);
     init_enumerate_device(info);
-    init_device(info);
     get_window_size(&info.width, &info.height);
     init_connection(info);
     init_window(info);
