+++ API-Samples/spirv_specialization/spirv_specialization.cpp
@@ -230,7 +230,7 @@ const std::string fragmentSPIRV_specialized =
 
 
 
-int main(int argc, char **argv)
+int sample_main()
 {
     VkResult U_ASSERT_ONLY res;
     struct sample_info info = {};
@@ -242,8 +242,7 @@ int main(int argc, char **argv)
     init_device_extension_names(info);
     init_instance(info, sample_title);
     init_enumerate_device(info);
-    init_device(info);
-    info.width = info.height = 500;
+    get_window_size(&info.width, &info.height);
     init_connection(info);
     init_window(info);
     init_swapchain_extension(info);
@@ -401,7 +400,7 @@ int main(int argc, char **argv)
     res = vkQueuePresentKHR(info.queue, &present);
     assert(res == VK_SUCCESS);
 
-    wait_seconds(1);
+    while(!wait(1)) {};
 
     vkDestroyFence(info.device, drawFence, NULL);
     vkDestroySemaphore(info.device, info.presentCompleteSemaphore, NULL);
