+++ API-Samples/utils/util_init.cpp
@@ -34,18 +34,6 @@ samples "init" utility functions
 #include "util_init.hpp"
 #include "cube_data.h"
 
-#ifdef __ANDROID__
-// Android specific stuff.
-
-// Header files..
-#include <android_native_app_glue.h>
-// Forward decl.
-bool Android_LoadFile(const char* filePath, std::vector<unsigned int>& data);
-
-// Static variable that keeps ANativeWindow and asset manager instances.
-static android_app* Android_application = nullptr;
-#endif
-
 using namespace std;
 
 /*
@@ -738,7 +726,7 @@ void init_swapchain_extension(struct sample_info &info)
     createInfo.sType = VK_STRUCTURE_TYPE_ANDROID_SURFACE_CREATE_INFO_KHR;
     createInfo.pNext = nullptr;
     createInfo.flags = 0;
-    createInfo.window = Android_application->window;
+    createInfo.window = AndroidGetApplicationWindow();
     res = info.fpCreateAndroidSurfaceKHR(info.inst, &createInfo, nullptr, &info.surface);
 #else  // !__ANDROID__ && !_WIN32
     VkXcbSurfaceCreateInfoKHR createInfo = {};
@@ -1528,12 +1516,13 @@ void init_shaders(struct sample_info &info, const char *vertShaderText, const ch
         info.shaderStages[0].pName = "main";
 
 #ifdef __ANDROID__
-        retVal = Android_LoadFile(vertShaderText, vtx_spv);
+        retVal = AndroidLoadFile(vertShaderText, vtx_spv);
 #else
         retVal = GLSLtoSPV(VK_SHADER_STAGE_VERTEX_BIT, vertShaderText, vtx_spv);
 #endif
         assert(retVal);
 
+        VkShaderModuleCreateInfo moduleCreateInfo;
         moduleCreateInfo.sType = VK_STRUCTURE_TYPE_SHADER_MODULE_CREATE_INFO;
         moduleCreateInfo.pNext = NULL;
         moduleCreateInfo.flags = 0;
@@ -1553,12 +1542,13 @@ void init_shaders(struct sample_info &info, const char *vertShaderText, const ch
         info.shaderStages[1].pName = "main";
 
 #ifdef __ANDROID__
-        retVal = Android_LoadFile(fragShaderText, frag_spv);
+        retVal = AndroidLoadFile(fragShaderText, frag_spv);
 #else
         retVal = GLSLtoSPV(VK_SHADER_STAGE_FRAGMENT_BIT, fragShaderText, frag_spv);
 #endif
         assert(retVal);
 
+        VkShaderModuleCreateInfo moduleCreateInfo;
         moduleCreateInfo.sType = VK_STRUCTURE_TYPE_SHADER_MODULE_CREATE_INFO;
         moduleCreateInfo.pNext = NULL;
         moduleCreateInfo.flags = 0;
@@ -2209,95 +2199,4 @@ void destroy_textures(struct sample_info &info)
         vkDestroyImage(info.device, info.textures[i].image, NULL);
         vkFreeMemory(info.device, info.textures[i].mem, NULL);
     }
-}
-
-bool get_window_size(int32_t* width, int32_t* height) {
-#ifndef __ANDROID__
-    // On Other platforms, set the window size.
-    *width = *height = 500;
-#else
-    // On Android, retrieve the window size from the native window.
-    assert(Android_application != nullptr);
-    *width = ANativeWindow_getWidth(Android_application->window);
-    *height = ANativeWindow_getHeight(Android_application->window);
-#endif
-    return true;
-};
-
-
-#ifndef __ANDROID__
-int main(int argc, char **argv) {
-    return sample_main();
-}
-#else
-//
-// Android specific helper functions.
-//
-void Android_handle_cmd(android_app* app, int32_t cmd)  {
-    switch( cmd ){
-        case APP_CMD_INIT_WINDOW:
-            // The window is being shown, get it ready.
-            sample_main();
-            break;
-        case APP_CMD_TERM_WINDOW:
-            // The window is being hidden or closed, clean it up.
-            break;
-        default :
-            LOGI("event not handled: %d", cmd);
-    }
-}
-
-bool Android_process_command() {
-    assert(Android_application != nullptr);
-    int events;
-    android_poll_source* source;
-    // Poll all pending events.
-    if( ALooper_pollAll(0, NULL, &events, (void**)&source) >= 0 ){
-        // Process each polled events
-        if (source != NULL)
-            source->process(Android_application, source);
-    }
-    return Android_application->destroyRequested;
-}
-
-void android_main(struct android_app* app) {
-    // Magic call, please ignore it (Android specific).
-    app_dummy();
-    // Set static variables.
-    Android_application = app;
-    // Set the callback to process system events
-    app->onAppCmd = Android_handle_cmd;
-
-    // Main loop
-    do {
-        Android_process_command();
-    } // Check if system requested to quit the application
-    while(app->destroyRequested == 0);
-
-    return;
-}
-
-bool Android_LoadFile(const char* filePath, std::vector<unsigned int> &data){
-    assert(Android_application != nullptr);
-    AAsset* file = AAssetManager_open(Android_application->activity->assetManager,
-                                      filePath, AASSET_MODE_BUFFER);
-    size_t fileLength = AAsset_getLength(file);
-    LOGI("Loaded file:%s size:%d", filePath, fileLength);
-    if (fileLength == 0) {
-        return false;
-    }
-    data.resize((fileLength + sizeof(uint32_t) - 1)/ sizeof(uint32_t));
-    AAsset_read(file, &data[0], fileLength);
-    return true;
-}
-#endif
-
-bool wait(int32_t timeout) {
-#ifndef __ANDROID__
-    wait_second(timeout);
-    return true;
-#else
-    (void)timeout;
-    return Android_process_command();
-#endif
 }
